backend/requirements.txt
fastapi[standard]
pydantic
pydantic-settings
google-generativeai
firebase-admin
python-dotenv
uvicorn
apscheduler
